+++ packages/pwa-kit-react-sdk/src/ssr/universal/components/refetch-data/index.test.js
@@ -8,7 +8,7 @@ import {useQueryClient} from '@tanstack/react-query'
 import {render, screen, waitFor} from '@testing-library/react'
 import React from 'react'
 import {useHistory, useLocation} from 'react-router-dom'
-import RefetchData from './refetch-data'
+import RefetchData from './index'
 
 jest.useFakeTimers()
